* removed unnecessary "dispose()" method in message encoder classes * added MsgError to protocol * fixed some typos * added missing @override annotations in message decoder classes * changed abstract message encoder class: message tspe is now read from message that is being sent, instead of providing the type in each encoder class again * a more detailed exception is thrown in case of a communication exception. Issue #82 is provisionally done.

+++ src/main/java/de/root1/simon/codec/messages/AbstractMessage.java
@@ -24,6 +24,10 @@ import de.root1.simon.Statics;
 
 /**
  * A base message for SIMON protocol messages.
+ * Error messages are not being sent via the errorMsg field:
+ * In case of a problems reading a message, the responsible decoder class will 
+ * inject the error message, so that the ProcessMessageRunnable can react on 
+ * that error accordingly.
  *
  * @author ACHR
  */
@@ -75,7 +79,7 @@ public abstract class AbstractMessage implements Serializable {
      * Returns the message type as described by {@link SimonMessageConstants}
      * @return the msgType
      */
-    public int getMsgType() {
+    public byte getMsgType() {
         return msgType;
     }
